sms/backend/requirements.txt
# FastAPI and server
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.0
pydantic-settings==2.2.0
email-validator==2.1.0.post1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.28
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0

# Utils
python-dotenv==1.0.1
httpx==0.27.0
tenacity==8.2.3
Jinja2==3.1.3
pyyaml==6.0.1
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.5
pytest-cov==4.1.0
faker==22.0.0 